+++ docs/changelog.md
@@ -4,6 +4,7 @@
 - Breaking change: for the `inject` command, change `--include-binaries` to `--include-apps`
 - Change all terminology from `binary` to `app` or `application`
 - Improve argument parsing for `pipx run` and `pipx runpip`
+- If `--force` is passed, remove existing files in PIPX_BIN_DIR
 
 0.13.2.3